RSpec tests for Rugged::Repository additions
To make it easier to test repository operations, this commit introduces
some RSpec helpers, matchers and context definitions. Central element
is the "with repository" shared context, which can be used in RSpec
definitions. It sets up a repository instance in `tmp/test.git` from
the tarball `spec/support/rugged/test.git.tar.gz`.

Any future RSpecs can use this repository. If any tags, commits, etc.
are needed for the spec, just add them to the extracted tarball
repository and rebuild the tarball.

Also this commit fixes a small error, when no refspecs are given to
`#walker_with` and all methods using it. It now correctly adds the
reference `HEAD` to the list.